requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
environ==1.0
environs==9.3.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
marshmallow==3.12.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.0.0

pycparser==2.20
PyJWT==2.1.0
pylint==2.6.0
pyparsing==2.4.7
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
